Add CUDA runtime support for op_mat
This makes op_decl_sparsity and op_decl_mat do the right thing on
CUDA.  Note that the sparsity is constructed on the host and then
moved to the device.  The matrix gets pointers to the device row and
column indices and allocates space for data entries.  The matrix is
stored in CSR format.