Add sequential support for op_iteration_space in parallel loops
The op_iteration_space construct allows a user to define the iteration
space of each element of a set.  At the moment, this can only be used
in parallel loops containing op_mat arguments.  An example calling
convention is:

  void kernel(float *A, float *x[2], int i, int j);

  op_par_loop(kernel, "kernel", op_iteration_space(set, 3, 3),
              op_arg_mat(mat, op_i(1), map, op_i(2), map, 1, "float", OP_INC),
              op_arg_dat(x, OP_ALL, map, 2, "float", OP_READ));

The kernel is now executed 9 times for each element in the set, with i
and j both running from 0 to 2:

  for ( int n = 0; n < set->size; n++ )
    for ( int i = 0; i < 3; i++ )
      for ( int j = 0; j < 3; j++ )
        kernel(&entry, data[n], i, j);

The magic symbol op_i(k) can be used to identify a map-index
argument with the corresponding kth argument of the op_iteration_space
(with the set argument numbered as 0).

So for example if we had two matrices, A and B, and were constructing
element blocks that were 3x4 for A and 4x3 for B, we might do
something like:

  op_par_loop(kernel, "kernel", op_iteration_space(set, 3, 4),
              op_arg_mat(A, op_i(1), map1, op_i(2), map2, "float", OP_INC),
              op_arg_mat(B, op_i(2), map3, op_i(1), map4, "float", OP_INC),
              op_arg_dat(x, OP_ALL, map, 2, "float", OP_READ));

The kernel knows that its arguments i,j construct the entry A[i][j],
but the entry B[j][i] and the op_i constructs in op_arg_mat ensure
that these entries are placed in the correct place in the global matrix.